requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
ultralytics>=8.3.0
opencv-python>=4.8.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
httpx==0.25.2
Pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0